Move JPAUrl to model subpackage
