+++ server/data/data-jpa/src/main/java/org/apache/james/mailrepository/jpa/model/JPAUrl.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailrepository.jpa;
+package org.apache.james.mailrepository.jpa.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
